Move UserModSelectOverlay to RoomSubScreen for Playlists consumption
