+++ osu.Game/Screens/OnlinePlay/Multiplayer/MultiplayerMatchSubScreen.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Diagnostics;
 using System.Linq;
 using JetBrains.Annotations;
@@ -16,7 +15,6 @@ using osu.Framework.Threading;
 using osu.Game.Configuration;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
-using osu.Game.Overlays.Mods;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Screens.OnlinePlay.Components;
 using osu.Game.Screens.OnlinePlay.Match;
@@ -43,9 +41,7 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
         [Resolved]
         private OngoingOperationTracker ongoingOperationTracker { get; set; }
 
-        private ModSelectOverlay userModsSelectOverlay;
         private MultiplayerMatchSettingsOverlay settingsOverlay;
-        private Drawable userModsSection;
 
         private IBindable<bool> isConnected;
 
@@ -155,7 +151,7 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
                                                                             new BeatmapSelectionControl { RelativeSizeAxes = Axes.X }
                                                                         }
                                                                     },
-                                                                    userModsSection = new FillFlowContainer
+                                                                    UserModsSection = new FillFlowContainer
                                                                     {
                                                                         RelativeSizeAxes = Axes.X,
                                                                         AutoSizeAxes = Axes.Y,
@@ -176,7 +172,7 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
                                                                                         Origin = Anchor.CentreLeft,
                                                                                         Width = 90,
                                                                                         Text = "Select",
-                                                                                        Action = () => userModsSelectOverlay.Show()
+                                                                                        Action = ShowUserModSelect,
                                                                                     },
                                                                                     new ModDisplay
                                                                                     {
@@ -231,19 +227,6 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
                         new Dimension(GridSizeMode.AutoSize),
                     }
                 },
-                new Container
-                {
-                    Anchor = Anchor.BottomLeft,
-                    Origin = Anchor.BottomLeft,
-                    RelativeSizeAxes = Axes.Both,
-                    Height = 0.5f,
-                    Padding = new MarginPadding { Horizontal = HORIZONTAL_OVERFLOW_PADDING },
-                    Child = userModsSelectOverlay = new UserModSelectOverlay
-                    {
-                        SelectedMods = { BindTarget = UserMods },
-                        IsValidMod = _ => false
-                    }
-                },
                 settingsOverlay = new MultiplayerMatchSettingsOverlay
                 {
                     RelativeSizeAxes = Axes.Both,
@@ -269,7 +252,6 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
         {
             base.LoadComplete();
 
-            Playlist.BindCollectionChanged(onPlaylistChanged, true);
             BeatmapAvailability.BindValueChanged(updateBeatmapAvailability, true);
             UserMods.BindValueChanged(onUserModsChanged);
 
@@ -303,30 +285,7 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
                 return true;
             }
 
-            if (userModsSelectOverlay.State.Value == Visibility.Visible)
-            {
-                userModsSelectOverlay.Hide();
-                return true;
-            }
-
             return base.OnBackButton();
-        }
-
-        private void onPlaylistChanged(object sender, NotifyCollectionChangedEventArgs e)
-        {
-            SelectedItem.Value = Playlist.FirstOrDefault();
-
-            if (SelectedItem.Value?.AllowedMods.Any() != true)
-            {
-                userModsSection.Hide();
-                userModsSelectOverlay.Hide();
-                userModsSelectOverlay.IsValidMod = _ => false;
-            }
-            else
-            {
-                userModsSection.Show();
-                userModsSelectOverlay.IsValidMod = m => SelectedItem.Value.AllowedMods.Any(a => a.GetType() == m.GetType());
-            }
         }
 
         private ModSettingChangeTracker modSettingChangeTracker;
@@ -432,10 +391,6 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer
             }
 
             modSettingChangeTracker?.Dispose();
-        }
-
-        private class UserModSelectOverlay : LocalPlayerModSelectOverlay
-        {
         }
     }
 }
